Modifying program parser for Bubblegum mint to do nothing on conflict
- Originally the `INSERT`s were changed to Upserts (use `INSERT` with
`ON CONFLICT` and using `>= asset.seq`), but I could not think of a
valid use case where mint instructions should be reran if the asset
already exists.
- So I changed it to not modify any existing values:
`ON CONFLICT DO NOTHING`.
- With this change, if a Bubblegum Mint instruction is reran it will
only create entries that don't exist.  The use case for this is if
a Mint instruction was only partially processed, for example due
to database errors.